+++ src/Ana/Ana/Ana.interface.ts
@@ -2,6 +2,8 @@
  * @module Ana/Interface
  */
 
+import { GenericData } from "../types"
+
 //  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
 //      _                  ___       _             __                
 //     / \   _ __   __ _  |_ _|_ __ | |_ ___ _ __ / _| __ _  ___ ___ 
@@ -11,23 +13,33 @@
 //
 ////////////////////////////////////////////////////////////////////////////////////////////////////////
 export interface iAnaConfiguration {
-  // includeComponentLibrary?: boolean
-  // includeAtoms?: boolean
-  // includeMolecules?: boolean
-  // includeOrganisms?: boolean
-  // includePages?: boolean
-  // includeSVGLibrary?: boolean
-  extensions?: { [key:string]: any }
+  /**
+   * This dictionary is used to extend the framework using official plugins.
+   */
+  extensions?: GenericData
 
-  // Component/Organisms/App
-  // appContainerId?: string = 'app' // index.html -> <body><div id="app"></div></body>
+  /**
+   * This is the list of supported SVG element tagnames.
+   */
+  svgElements?: string[]
+
+  /**
+   * This is the list of HTMLElement tagnames marked as empty. 
+   */
+  emptyElements?: string[]
+
+  /**
+   * index.html -> <body><div id="app"></div></body>
+   */
+  appContainerId?: string
 }
 
 //  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
 export interface AnaConfiguration {
-  extensions: { [key:string]: any }
+  extensions: GenericData
   svgElements: string[]
   emptyElements: string[]
+  appContainerId: string
 }
 
 //  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
@@ -37,4 +49,5 @@ export const dAnaConfiguration: AnaConfiguration = {
   svgElements: ['circle','ellipse','line','polygon','polyline','rect','defs','g','marker','mask','svg','switch','symbol','desc','linearGradient','radialGradient','stop','image','path','text','use'],
   // prettier-ignore
   emptyElements: ['area','base','br','col','embed','hr','img','input','link','meta','param','source','track','wbr'],
+  appContainerId: 'app'
 }