+++ src/Ana/Ana/Ana.ts
@@ -6,21 +6,8 @@ import {
   dAnaConfiguration,
   iAnaConfiguration,
 } from './Ana.interface'
-import { GenericData, AttributeValuesDictionary } from '../types'
-import { Observable } from '../Observable/Observable'
-import { createRenderer } from '../Render/Render'
 import { applyDefaultParameters } from '../Utils/Utils'
-import { Render } from '../Render/Render.interface'
-import { App } from '../App/App'
-
-declare global {
-  interface HTMLElement {
-    has(attributes: AttributeValuesDictionary): HTMLElement
-  }
-  interface SVGElement {
-    has(attributes: AttributeValuesDictionary): HTMLElement
-  }
-}
+import { ReactiveRenderer } from '../Render/Render'
 
 //  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
 //       _
@@ -35,91 +22,24 @@ declare global {
  */
 export class Ana {
   /**
-   * 
+   *
    */
-  createApp: App = new App()
-
-  /**
-   * This private observable is in charge of making the UI react to changes in the state.
-   */
-  private obs: Observable = new Observable()
-
-  /**
-   * This object functions as storage for the state up to the latest changes.
-   */
-  private state: GenericData = {}
-
-  /**
-   * This function creates the starting app for the responsive rendering of a page.
-   */
-  app(state: GenericData, render: Function): void {
-    this.obs.subscribe(render)
-    this.up(state)
-  }
-
-  /**
-   * This function updates the state and emits the change to the observable.
-   */
-  up(state: GenericData) {
-    this.state = { ...this.state, ...state }
-    this.obs.emit(this.state)
-  }
-
-  /**
-   * This function creates a dictionary of render functions for HTMLElements and UI components from the Ana framework.
-   */
-  render: Render
+  createApp: ReactiveRenderer
 
   /**
    * This function instantiates the framework.
    */
   constructor(configuration: iAnaConfiguration = {}) {
-    let config = applyDefaultParameters<
-      AnaConfiguration,
-      iAnaConfiguration
-    >(dAnaConfiguration, configuration)
+    let config = applyDefaultParameters<AnaConfiguration, iAnaConfiguration>(
+      dAnaConfiguration,
+      configuration
+    )
 
-    HTMLElement.prototype.has = has
-    SVGElement.prototype.has = has
+    this.createApp = new ReactiveRenderer(config)
 
     // Adds ana.js-check
-    if(config.extensions.check) {
+    if (config.extensions.check) {
       config.extensions.check()
     }
-
-    // Adds ana.js-ui
-    if(config.extensions.atoms) {
-      this.render = config.extensions.atoms
-    } else {
-      this.render = createRenderer(config)
-    }
   }
-}
-
-//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
-/**
- * This function extends HTMLElement.prototype.setAttribute to support a dictionary of attributes instead of setting them one by one.
- */
-export const has = function (
-  this: HTMLElement,
-  attributes: AttributeValuesDictionary
-): HTMLElement {
-  Object.keys(attributes).forEach((attributeName: string) => {
-    let attribute = attributes[attributeName]
-    if (typeof attribute === 'string') {
-      this.setAttribute(attributeName, attribute)
-    } else if (typeof attribute === 'boolean') {
-      if (attribute === true) {
-        this.setAttribute(attributeName, '')
-      } else {
-        this.removeAttribute(attributeName)
-      }
-    } else {
-      let listenerFunction: Function = attribute
-      this.addEventListener(attributeName, (event: Event) => {
-        listenerFunction(event)
-      })
-    }
-  })
-  return this
 }