+++ src/main.ts
@@ -1,14 +1,38 @@
-import { Ana, byId } from './Ana/index'
+import { Ana, GenericData } from './Ana/index'
 
-const A = new Ana()
-const a = A.render
+const ana = new Ana()
+const app = ana.createApp
+const a = app.render
 
-A.app({
-  title: 'Example title'
-}, (state: any) => byId('app').replaceChildren(
-  a.h1()(state.title)
-))
+app.init(
+  {
+    a: 1,
+    b: 2,
+    c: 3,
+    d: 4,
+    e: 5,
+  },
+  (d: GenericData) =>
+    a.ol()(a.li()(d.a), a.li()(d.b), a.li()(d.c), a.li()(d.d), a.li()(d.e))
+)
 
 setTimeout(() => {
-  A.up({ title: 'Title changed after 5 seconds' })
-}, 5000);
+  renderRandoms()
+}, 1000)
+
+var interval: NodeJS.Timeout | undefined = undefined
+
+async function renderRandoms() {
+  let options: string[] = ['a', 'b', 'c', 'd', 'e']
+  interval = setInterval(() => {
+    let updateSnapshot: GenericData = {}
+    updateSnapshot[options[Math.floor(Math.random() * 5)]] = Math.floor(
+      Math.random() * 100
+    )
+    app.up(updateSnapshot)
+  }, 1000)
+}
+
+document.body.addEventListener('click', () => {
+  if (interval !== undefined) clearInterval(interval)
+})